backend/requirements.txt
Django==5.2.1                      
djangorestframework==3.16.0        
djoser==2.3.1                      
djangorestframework-simplejwt==5.5.0 
django-filter==24.2
psycopg2-binary==2.9.9
Pillow==10.3.0                     
python-dotenv==1.0.1               

gunicorn==23.0.0                    
drf-spectacular==0.27.2          
drf-nested-routers==0.93.4         

flake8==7.0.0
isort==5.13.2
pytest==8.2.0
pytest-django==4.8.0
coverage==7.5.1
